+++ src/main/java/com/BankingApplication/BankApplication/dao/DebitCardDao.java
@@ -1,14 +1,11 @@
 package com.BankingApplication.BankApplication.dao;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import com.BankingApplication.BankApplication.entity.Account;
 import com.BankingApplication.BankApplication.entity.DebitCard;
 import com.BankingApplication.BankApplication.entity.Status;
-import com.BankingApplication.BankApplication.exception.AccountNotFoundException;
 import com.BankingApplication.BankApplication.exception.DebitCardNotFoundException;
 import com.BankingApplication.BankApplication.repository.DebitCardRepository;
 
@@ -35,13 +32,13 @@ public class DebitCardDao {
 	}
 
 	// update name
-	public DebitCard getDebitCardByNumber(int cardNumber) {
+	public DebitCard getDebitCardByNumber(long cardNumber) {
 		return cardRepository.findById(cardNumber)
 				.orElseThrow(() -> new DebitCardNotFoundException("Card does not exist"));
 	}
 
 	// update name
-	public DebitCard updateName(int cardNumber, String name) {
+	public DebitCard updateName(long cardNumber, String name) {
 		DebitCard card = getDebitCardByNumber(cardNumber);
 		card.setName(name);
 		saveDebitCard(card);
@@ -49,7 +46,7 @@ public class DebitCardDao {
 	}
 
 	// update phone
-	public DebitCard updatePhone(int cardNumber, long phone) {
+	public DebitCard updatePhone(long cardNumber, long phone) {
 		DebitCard card = getDebitCardByNumber(cardNumber);
 		card.setPhone(phone);
 		saveDebitCard(card);
@@ -57,7 +54,7 @@ public class DebitCardDao {
 	}
 
 	// update status
-	public DebitCard updateStatus(int cardNumber, Status status) {
+	public DebitCard updateStatus(long cardNumber, Status status) {
 		DebitCard card = getDebitCardByNumber(cardNumber);
 		card.setStatus(status);
 		saveDebitCard(card);
@@ -65,7 +62,7 @@ public class DebitCardDao {
 	}
 
 	// delete card
-	public String deleteDebitCard(int cardNumber) {
+	public String deleteDebitCard(long cardNumber) {
 		DebitCard card = getDebitCardByNumber(cardNumber);
 		Account account = card.getAccount();
 		if (account != null) {
