+++ src/main/java/com/BankingApplication/BankApplication/service/DebitCardService.java
@@ -5,7 +5,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import com.BankingApplication.BankApplication.dao.AccountDao;
 import com.BankingApplication.BankApplication.dao.DebitCardDao;
 import com.BankingApplication.BankApplication.entity.DebitCard;
 import com.BankingApplication.BankApplication.entity.Status;
@@ -28,7 +27,7 @@ public class DebitCardService {
 	}
 
 	// get by card number
-	public ResponseEntity<ResponseStructure<DebitCard>> getCardByCardNumber(int cardNumber) {
+	public ResponseEntity<ResponseStructure<DebitCard>> getCardByCardNumber(long cardNumber) {
 		DebitCard card2 = cardDao.getDebitCardByNumber(cardNumber);
 		ResponseStructure<DebitCard> responseStructure = new ResponseStructure<>();
 		responseStructure.setData(card2);
@@ -39,7 +38,7 @@ public class DebitCardService {
 	}
 
 	// update card name
-	public ResponseEntity<ResponseStructure<DebitCard>> updateName(int cardNumber, String name) {
+	public ResponseEntity<ResponseStructure<DebitCard>> updateName(long cardNumber, String name) {
 		DebitCard card2 = cardDao.updateName(cardNumber, name);
 		ResponseStructure<DebitCard> responseStructure = new ResponseStructure<>();
 		responseStructure.setData(card2);
@@ -49,7 +48,7 @@ public class DebitCardService {
 	}
 
 	// update card phone
-	public ResponseEntity<ResponseStructure<DebitCard>> updatePhone(int cardNumber, long phone) {
+	public ResponseEntity<ResponseStructure<DebitCard>> updatePhone(long cardNumber, long phone) {
 		DebitCard card2 = cardDao.updatePhone(cardNumber, phone);
 		ResponseStructure<DebitCard> responseStructure = new ResponseStructure<>();
 		responseStructure.setData(card2);
@@ -59,7 +58,7 @@ public class DebitCardService {
 	}
 
 	// update card status
-	public ResponseEntity<ResponseStructure<DebitCard>> updateStatus(int cardNumber, Status status) {
+	public ResponseEntity<ResponseStructure<DebitCard>> updateStatus(long cardNumber, Status status) {
 		DebitCard card2 = cardDao.updateStatus(cardNumber, status);
 		ResponseStructure<DebitCard> responseStructure = new ResponseStructure<>();
 		responseStructure.setData(card2);
@@ -69,7 +68,7 @@ public class DebitCardService {
 	}
 	
 	// delete card by number
-	public ResponseEntity<ResponseStructure<String>> deleteDebitCard(int cardNumber) {
+	public ResponseEntity<ResponseStructure<String>> deleteDebitCard(long cardNumber) {
 		String response = cardDao.deleteDebitCard(cardNumber);
 		ResponseStructure<String> responseStructure = new ResponseStructure<>();
 		responseStructure.setData(response);
